LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/eval-allowed-in-report-only-mode-and-sends-report-expected.txt


PASS Eval is allowed because the CSP is report-only
FAIL Violation report status OK. assert_true: No CSP report sent, but expecting one. expected true got false
